# docs/swagger.yaml
basePath: /
definitions:
...
  dto.MessageRequestQuoted:
    properties:
      key:
        $ref: '#/definitions/dto.QuotedKey'
      message:
        $ref: '#/definitions/dto.QuotedMessage'
    type: object
...
  dto.SendAudioRequest:
    properties:
      audio:
        type: string
      delay:
        maximum: 300000
        minimum: 0
        type: integer
      encoding:
        type: boolean
      mentioned:
        items:
          type: string
        type: array
      mentionsEveryOne:
        type: boolean
      number:
        type: string
      quoted:
        $ref: '#/definitions/dto.MessageRequestQuoted'
    type: object
  dto.SendAudioResponse:
    properties:
      contextInfo:
        $ref: '#/definitions/dto.MessageContextInfo'
      instanceId:
        type: string
      key:
        $ref: '#/definitions/dto.MessageResponseKey'
      message:
        $ref: '#/definitions/dto.SendAudioResponseMessage'
      messageTimestamp:
        type: integer
      messageType:
        type: string
      pushName:
        type: string
      source:
        type: string
      status:
        type: string
    type: object
...
  utils.AuthenticationErrorResponse:
    properties:
      message:
        example: Unauthorized
        type: string
    type: object
  utils.HTTPErrorResponse:
    properties:
      error:
        description: |-
          Error is a legacy opaque value. Go's JSON encoder does not serialize an
          error message here, so clients should use Message and ErrorMessage.
        type: object
      errorMessage:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Verbeux AI
    url: https://github.com/verbeux-ai/whatsmiau
  description: Authenticated WhatsApp API compatible with Evolution API. The specification
    covers the complete application route surface, including canonical routes, compatibility
    aliases, calls and documentation delivery.
  license:
    name: MIT
  title: WhatsMiau API
  version: 0.3.2
paths:
...
  /v1/instance/{instance}/message/audio:
    post:
      consumes:
      - application/json
      description: Sends an audio file (by URL) as a WhatsApp voice message to the
        specified number
      operationId: postV1InstanceByInstanceMessageAudio
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Audio message parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendAudioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendAudioResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send an audio message
      tags:
      - Message
  /v1/instance/{instance}/message/buttons:
    post:
      consumes:
      - application/json
      description: Sends an interactive buttons message (reply type) or PIX payment
      operationId: postV1InstanceByInstanceMessageButtons
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Buttons message parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendButtonsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendButtonsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send a buttons message
      tags:
      - Message
...
  /v1/instance/{instance}/message/document:
    post:
      consumes:
      - application/json
      description: Sends a document file by URL to a WhatsApp number
      operationId: postV1InstanceByInstanceMessageDocument
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Document parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendDocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendDocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send a document
      tags:
      - Message
  /v1/instance/{instance}/message/image:
    post:
      consumes:
      - application/json
      description: Sends an image file by URL to a WhatsApp number
      operationId: postV1InstanceByInstanceMessageImage
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Image parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendDocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendDocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send an image
      tags:
      - Message
  /v1/instance/{instance}/message/list:
    post:
      consumes:
      - application/json
      description: Sends an interactive list message with selectable options
      operationId: postV1InstanceByInstanceMessageList
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: List message parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send a list message
      tags:
      - Message
...
  /v1/instance/{instance}/message/text:
    post:
      consumes:
      - application/json
      description: Sends a text message to a WhatsApp number via the specified instance
      operationId: postV1InstanceByInstanceMessageText
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Text message parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendTextRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendTextResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send a text message
      tags:
      - Message
...
  /v1/message/sendButtons/{instance}:
    post:
      consumes:
      - application/json
      description: Sends an interactive buttons message (reply type) or PIX payment
      operationId: postV1MessageSendButtonsByInstance
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Buttons message parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendButtonsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendButtonsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send a buttons message
      tags:
      - Message
...
  /v1/message/sendList/{instance}:
    post:
      consumes:
      - application/json
      description: Sends an interactive list message with selectable options
      operationId: postV1MessageSendListByInstance
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: List message parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send a list message
      tags:
      - Message
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: apikey
    type: apiKey
swagger: "2.0"